refactor: extract utility functions and constants
- `getMdxFiles` 함수를 별도 utils 파일로 분리
- `BASE_URL` 및 `POSTS_DIR`을 constants 파일로 분리
