+++ src/app/api/posts/route.ts
@@ -1,28 +1,8 @@
-import { promises as fs } from 'fs';
 import { NextResponse } from 'next/server';
-import path from 'path';
 
+import { POSTS_DIR } from '@lib/api/constants';
+import { getMdxFiles } from '@lib/api/utils';
 import { PostFactory } from '@lib/posts/postFactory';
-
-const POSTS_DIR = path.join(process.cwd(), 'src/posts');
-
-async function getMdxFiles(dir: string): Promise<string[]> {
-  const entries = await fs.readdir(dir, { withFileTypes: true });
-  const files = await Promise.all(
-    entries.map(async entry => {
-      const fullPath = path.join(dir, entry.name);
-
-      if (entry.isDirectory()) {
-        return await getMdxFiles(fullPath);
-      } else if (entry.name.endsWith('.mdx')) {
-        return fullPath;
-      } else {
-        return [];
-      }
-    })
-  );
-  return files.flat();
-}
 
 export async function GET(request: Request) {
   try {
