+++ src/domains/post/list/_hooks/usePosts.ts
@@ -1,5 +1,6 @@
-import { useQuery } from '@tanstack/react-query';
+import { useSuspenseQuery } from '@tanstack/react-query';
 
+import { BASE_URL } from '@lib/api/constants';
 import { Post } from '@lib/posts/postFactory';
 
 export type PostsResponse = {
@@ -8,7 +9,7 @@ export type PostsResponse = {
 };
 
 async function fetchPosts(category: string): Promise<PostsResponse> {
-  const url = `/api/posts?category=${encodeURIComponent(category)}`;
+  const url = `${BASE_URL}/api/posts?category=${encodeURIComponent(category)}`;
   const res = await fetch(url, {
     headers: {
       'Content-Type': 'application/json',
@@ -22,9 +23,8 @@ async function fetchPosts(category: string): Promise<PostsResponse> {
 }
 
 export default function usePosts(selectedCategory: string) {
-  return useQuery({
+  return useSuspenseQuery({
     queryKey: ['posts', selectedCategory],
     queryFn: () => fetchPosts(selectedCategory),
-    staleTime: 1000 * 60 * 5,
   });
 }
